+++ src/main/java/com/moriba/skultem/application/usecase/CountStudentFeesUseCase.java
@@ -1,5 +1,7 @@
 package com.moriba.skultem.application.usecase;
 
+import java.math.BigDecimal;
+
 import org.springframework.stereotype.Service;
 
 import com.moriba.skultem.domain.repository.StudentFeeRepository;
@@ -12,7 +14,7 @@ public class CountStudentFeesUseCase {
 
     private final StudentFeeRepository repo;
 
-    public Double execute(String school, String studentId) {
+    public BigDecimal execute(String school, String studentId) {
         return repo.sumTotalFeeByStudent(studentId, school);
     }
 }
